refactor: loadGeoJson into own file
