+++ src/js/modules/debug.js
@@ -1,0 +1,9 @@
+import rk from "./base_radlkarte_object";
+
+function debug(obj) {
+  if (rk.debug) {
+    console.log(obj);
+  }
+}
+
+export default debug;
